+++ FWI/Python/daily_summaries.py
@@ -106,7 +106,7 @@ def generate_daily_summaries(hourly_FWI, reset_hr = 5,
   
   for stn, by_stn in hourly_data.groupby("id", sort = False):
     if not silent:
-      print("Summarizing " + stn + " to daily")
+      print("Summarizing " + str(stn) + " to daily")
     by_stn["pseudo_DATE"] = by_stn.apply(lambda row:
       pseudo_date(row["yr"], row["mon"], row["day"], row["hr"], reset_hr), axis = 1)
     
@@ -177,7 +177,7 @@ def generate_daily_summaries(hourly_FWI, reset_hr = 5,
   if not (round_out == None or round_out == "None"):
     outcols = ["ffmc", "dmc", "dc", "isi", "bui", "fwi", "dsr",
       "gfmc", "gsi", "gfwi", "ws_smooth", "isi_smooth", "gsi_smooth"]
-    results[outcols] = results[outcols].map(round, ndigits = round_out)
+    results[outcols] = results[outcols].map(round, ndigits = int(round_out))
     
   return results