+++ FWI/R/daily_summaries.r
@@ -96,7 +96,7 @@ pseudo_date <- function(yr, mon, day, hr, reset_hr = 5) {
 #' @param     hourly_FWI    hourly FWI dataframe (output of hFWI())
 #' @param     reset_hr      new boundary to define day to summarize (default 5)
 #' @param     silent        suppresses informative print statements (default False)
-#' @param     round_out     decimals to truncate output to, None for none (default 4)
+#' @param     round_out     decimals to truncate output to, NA for none (default 4)
 #' @return                  daily summary of peak FWI conditions
 generate_daily_summaries <- function(hourly_FWI, reset_hr = 5,
   silent = FALSE, round_out = 4) {
@@ -195,10 +195,10 @@ generate_daily_summaries <- function(hourly_FWI, reset_hr = 5,
   }
 
   # format decimal places of output columns
-  if (!is.na(round_out)) {
+  if (!(is.na(round_out) || round_out == "NA")) {
     outcols <- c("ffmc", "dmc", "dc", "isi", "bui", "fwi", "dsr",
       "gfmc", "gsi", "gfwi", "ws_smooth", "isi_smooth", "gsi_smooth")
-    set(results, j = outcols, value = round(results[, ..outcols], round_out))
+    set(results, j = outcols, value = round(results[, ..outcols], as.integer(round_out)))
   }
 
   if (wasDf) {
@@ -218,7 +218,7 @@ if ("--args" %in% commandArgs() && sys.nframe() == 0) {
   input <- args[1]
   output <- args[2]
   # load optional arguments if provided, or set to default
-  if (length(args) >= 3) reset_hr <- args[3]
+  if (length(args) >= 3) reset_hr <- as.integer(args[3])
   else reset_hr <- 5
   if (length(args) >= 4) silent <- as.logical(args[4])
   else silent <- FALSE